+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 ### Fixed
 - `write_html`: headings (`<h1>`, `<h2>`...) can now contain non-ASCII characters without triggering a `UnicodeEncodeError`
 - `Template`: CSV column types are now safely parsed, thanks to @gmischler
+- `cell(..., markdown=True)` "leaked" its final style (bold / italics / underline) onto the following cells
 ### Changed
 - `write_html`: the line height of headings (`<h1>`, `<h2>`...) is now properly scaled with its font size
 - some `FPDF` methods should not be used inside a `rotation` context, or things can get broken.